refactor: extract hooks tests to separate module for improved testability
Extract internal testing functions from hooks.rs into a public module
(hooks_tests.rs) to enable better test organization and maintainability.
This refactor exposes several internal functions with pub(crate) visibility
to support test cases:

- is_system_event_hook: Check if a hook is a system event
- truncate_output: Truncate output to a maximum number of lines
- collect_stop_commands: Collect stop commands from configuration
- match_subagent_patterns: Match agent patterns against glob patterns

Additional changes:
- Organize imports alphabetically in hooks.rs for consistency
- Update config_tests.rs with refactored test organization
- Sort imports in hooks_tests.rs for clarity
- Add new hooks_tests.rs module with extracted test utilities
- Update related files to support the new module structure

This change improves code organization without altering functionality or
the public API.
