+++ src/schema.rs
@@ -46,7 +46,7 @@
 
 use anyhow::{Context, Result};
 use schemars::schema_for;
-use serde_json::{Value, json};
+use serde_json::{json, Value};
 use std::fs;
 use std::path::PathBuf;
