+++ tests/hooks_tests.rs
@@ -125,15 +125,13 @@ fn test_is_root_addition_false_cases() {
 #[test]
 fn test_matches_uneditable_pattern() {
     // Exact file matches
-    assert!(
-        matches_uneditable_pattern(
-            "package.json",
-            "package.json",
-            "/path/package.json",
-            "package.json"
-        )
-        .unwrap()
-    );
+    assert!(matches_uneditable_pattern(
+        "package.json",
+        "package.json",
+        "/path/package.json",
+        "package.json"
+    )
+    .unwrap());
 
     // Wildcard matches
     assert!(matches_uneditable_pattern("test.md", "test.md", "/path/test.md", "*.md").unwrap());
@@ -142,38 +140,32 @@ fn test_matches_uneditable_pattern() {
     );
 
     // Directory pattern matches
-    assert!(
-        matches_uneditable_pattern(
-            "src/index.ts",
-            "src/index.ts",
-            "/path/src/index.ts",
-            "src/**/*.ts"
-        )
-        .unwrap()
-    );
-    assert!(
-        matches_uneditable_pattern(
-            "src/lib/utils.ts",
-            "src/lib/utils.ts",
-            "/path/src/lib/utils.ts",
-            "src/**/*.ts"
-        )
-        .unwrap()
-    );
+    assert!(matches_uneditable_pattern(
+        "src/index.ts",
+        "src/index.ts",
+        "/path/src/index.ts",
+        "src/**/*.ts"
+    )
+    .unwrap());
+    assert!(matches_uneditable_pattern(
+        "src/lib/utils.ts",
+        "src/lib/utils.ts",
+        "/path/src/lib/utils.ts",
+        "src/**/*.ts"
+    )
+    .unwrap());
 
     // Negative matches
     assert!(
         !matches_uneditable_pattern("other.txt", "other.txt", "/path/other.txt", "*.md").unwrap()
     );
-    assert!(
-        !matches_uneditable_pattern(
-            "lib/index.ts",
-            "lib/index.ts",
-            "/path/lib/index.ts",
-            "src/**/*.ts"
-        )
-        .unwrap()
-    );
+    assert!(!matches_uneditable_pattern(
+        "lib/index.ts",
+        "lib/index.ts",
+        "/path/lib/index.ts",
+        "src/**/*.ts"
+    )
+    .unwrap());
 }
 
 #[test]
@@ -185,33 +177,27 @@ fn test_matches_uneditable_pattern_invalid_glob() {
 #[test]
 fn test_matches_uneditable_pattern_multiple_patterns() {
     // Test multiple patterns separately (since the glob crate doesn't support brace expansion)
-    assert!(
-        matches_uneditable_pattern(
-            "package.json",
-            "package.json",
-            "/path/package.json",
-            "package.json"
-        )
-        .unwrap()
-    );
-    assert!(
-        matches_uneditable_pattern(
-            "tsconfig.json",
-            "tsconfig.json",
-            "/path/tsconfig.json",
-            "tsconfig.json"
-        )
-        .unwrap()
-    );
-    assert!(
-        !matches_uneditable_pattern(
-            "other.json",
-            "other.json",
-            "/path/other.json",
-            "package.json"
-        )
-        .unwrap()
-    );
+    assert!(matches_uneditable_pattern(
+        "package.json",
+        "package.json",
+        "/path/package.json",
+        "package.json"
+    )
+    .unwrap());
+    assert!(matches_uneditable_pattern(
+        "tsconfig.json",
+        "tsconfig.json",
+        "/path/tsconfig.json",
+        "tsconfig.json"
+    )
+    .unwrap());
+    assert!(!matches_uneditable_pattern(
+        "other.json",
+        "other.json",
+        "/path/other.json",
+        "package.json"
+    )
+    .unwrap());
 }
 
 #[test]
@@ -221,51 +207,46 @@ fn test_matches_uneditable_pattern_environment_files() {
         matches_uneditable_pattern(".env.local", ".env.local", "/path/.env.local", ".env*")
             .unwrap()
     );
-    assert!(
-        matches_uneditable_pattern(
-            ".env.production",
-            ".env.production",
-            "/path/.env.production",
-            ".env*"
-        )
-        .unwrap()
-    );
-    assert!(
-        !matches_uneditable_pattern(
-            "environment.txt",
-            "environment.txt",
-            "/path/environment.txt",
-            ".env*"
-        )
-        .unwrap()
-    );
+    assert!(matches_uneditable_pattern(
+        ".env.production",
+        ".env.production",
+        "/path/.env.production",
+        ".env*"
+    )
+    .unwrap());
+    assert!(!matches_uneditable_pattern(
+        "environment.txt",
+        "environment.txt",
+        "/path/environment.txt",
+        ".env*"
+    )
+    .unwrap());
 }
 
 #[test]
 fn test_matches_uneditable_pattern_directory_patterns() {
     // Match entire directories
-    assert!(
-        matches_uneditable_pattern(
-            "docs/README.md",
-            "docs/README.md",
-            "/path/docs/README.md",
-            "docs/**"
-        )
-        .unwrap()
-    );
-    assert!(
-        matches_uneditable_pattern(
-            "docs/api/index.md",
-            "docs/api/index.md",
-            "/path/docs/api/index.md",
-            "docs/**"
-        )
-        .unwrap()
-    );
-    assert!(
-        !matches_uneditable_pattern("src/docs.ts", "src/docs.ts", "/path/src/docs.ts", "docs/**")
-            .unwrap()
-    );
+    assert!(matches_uneditable_pattern(
+        "docs/README.md",
+        "docs/README.md",
+        "/path/docs/README.md",
+        "docs/**"
+    )
+    .unwrap());
+    assert!(matches_uneditable_pattern(
+        "docs/api/index.md",
+        "docs/api/index.md",
+        "/path/docs/api/index.md",
+        "docs/**"
+    )
+    .unwrap());
+    assert!(!matches_uneditable_pattern(
+        "src/docs.ts",
+        "src/docs.ts",
+        "/path/src/docs.ts",
+        "docs/**"
+    )
+    .unwrap());
 }
 
 // Integration test for path normalization scenarios
@@ -1122,11 +1103,9 @@ fn test_validate_subagent_stop_payload_missing_agent_transcript_path() {
         result.is_err(),
         "Payload with empty agent_transcript_path should fail validation"
     );
-    assert!(
-        result
-            .unwrap_err()
-            .contains("agent_transcript_path cannot be empty")
-    );
+    assert!(result
+        .unwrap_err()
+        .contains("agent_transcript_path cannot be empty"));
 }
 
 #[test]
@@ -1139,11 +1118,9 @@ fn test_validate_subagent_stop_payload_whitespace_agent_transcript_path() {
         result.is_err(),
         "Payload with whitespace-only agent_transcript_path should fail validation"
     );
-    assert!(
-        result
-            .unwrap_err()
-            .contains("agent_transcript_path cannot be empty")
-    );
+    assert!(result
+        .unwrap_err()
+        .contains("agent_transcript_path cannot be empty"));
 }
 
 #[test]
@@ -1724,11 +1701,9 @@ fn test_validate_subagent_start_payload_missing_subagent_type() {
         result.is_err(),
         "Payload with empty subagent_type should fail validation"
     );
-    assert!(
-        result
-            .unwrap_err()
-            .contains("subagent_type cannot be empty")
-    );
+    assert!(result
+        .unwrap_err()
+        .contains("subagent_type cannot be empty"));
 }
 
 #[test]
@@ -1741,11 +1716,9 @@ fn test_validate_subagent_start_payload_whitespace_subagent_type() {
         result.is_err(),
         "Payload with whitespace-only subagent_type should fail validation"
     );
-    assert!(
-        result
-            .unwrap_err()
-            .contains("subagent_type cannot be empty")
-    );
+    assert!(result
+        .unwrap_err()
+        .contains("subagent_type cannot be empty"));
 }
 
 #[test]
@@ -1758,11 +1731,9 @@ fn test_validate_subagent_start_payload_missing_agent_transcript_path() {
         result.is_err(),
         "Payload with empty agent_transcript_path should fail validation"
     );
-    assert!(
-        result
-            .unwrap_err()
-            .contains("agent_transcript_path cannot be empty")
-    );
+    assert!(result
+        .unwrap_err()
+        .contains("agent_transcript_path cannot be empty"));
 }
 
 #[test]
@@ -1775,11 +1746,9 @@ fn test_validate_subagent_start_payload_whitespace_agent_transcript_path() {
         result.is_err(),
         "Payload with whitespace-only agent_transcript_path should fail validation"
     );
-    assert!(
-        result
-            .unwrap_err()
-            .contains("agent_transcript_path cannot be empty")
-    );
+    assert!(result
+        .unwrap_err()
+        .contains("agent_transcript_path cannot be empty"));
 }
 
 #[test]
